Changes:
1. Updated coil to version 2.3.0.
2. Added removeFromDataBase function in Dao.kt.
3. Organised and simplified FavoriteScreen.kt.
4. Dropdown menu now has "delete" button if you already have anime in database.

+++ app/src/main/java/com/example/animeapp/domain/models/detailModel/Data.kt
@@ -35,7 +35,7 @@ data class Data(
     @SerializedName("rank") val rank: Int,
     @SerializedName("rating") val rating: String,
     @SerializedName("score") val score: Float,
-    @SerializedName("scored_by") val scored_by: Int,
+    @SerializedName("scored_by") val scored_by: Float,
     @SerializedName("season") val season: String,
     @SerializedName("source") val source: String,
     @SerializedName("status") val status: String,
